+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.2.3
 - The pasted text will undergo additional cleaning.
 - CSS refactor to make maintenance easier.
-- [Added] Class item export support in Statblock Exporter (domains, features, subclasses, inventory, background questions, connections)
+- [Added] Class item export support for class, ancestry and community
 
 # 0.2.2
 - Bug Fix for Consumable